third_party/WebKit/LayoutTests/scrollbars/auto-scrollbar-zero-width-block.html

<!DOCTYPE html>
<style>
#floated-container {
  float: left;
  height: 100px;
  overflow: auto;
}
#empty-block {
  height: 200px;
}
</style>
<div>
  <div id="floated-container">
    <div id="empty-block"></div>
  </div>
</div>
<script>
if (window.testRunner)
  testRunner.dumpAsText();

var result = document.createElement("p");
document.body.appendChild(result);
var container = document.getElementById("floated-container");
if (container.clientWidth == 0)
  result.innerHTML = "PASS: container.clientWidth is zero.";
else
  result.innerHTML = "FAIL: container.clientWidth is not zero.";
</script>
