third_party/blink/web_tests/scrollbars/auto-scrollbar-zero-width-block-expected.txt

PASS: container.clientWidth is zero.
